Logic/requirements.txt
gpiozero>=1.6.0
pyrplidar>=0.1.0
pyserial>=3.5
pygame>=2.0.0

